priprava 2.5.29
git-svn-id: https://svn.code.sf.net/p/epos/code/trunk@154 905ca500-da46-4645-b043-b538ca7892eb

+++ src/epos.h
@@ -88,7 +88,7 @@ typedef char UNIT;
 #define U_VOID		120
 
 extern int unused_variable;
-#define unuse(x) (unused_variable = (int)(x));
+#define unuse(x) (unused_variable = (long int)(x));
 
 extern const bool is_monolith;
